DATACOUCH-626 Handle all-args constructor with version being a simple type (long).
The issue was that on read, when an all-args constructor is used, the version
passed in is always null because it is taken from the cas and not the document,
and this fails as a null cannot be supplied for a simple type.
This fix pre-populates the 'converted' document with the cas, just as already
done with the id.